fix: theme treated as Context
