+++ template/src/components/App/App.js
@@ -8,18 +8,17 @@ import { CssBaseline, Button, Snackbar } from "@mui/material";
 
 import { auth, firestore } from "../../firebase";
 import authentication from "../../services/authentication";
-import appearance from "../../services/appearance";
 
 import ErrorBoundary from "../ErrorBoundary";
 import LaunchScreen from "../LaunchScreen";
 import Bar from "../Bar";
 import Router from "../Router";
 import DialogHost from "../DialogHost";
+import { appearance, ThemeContext } from "../../context/ThemeContext";
 
 const initialState = {
   ready: false,
   performingAction: false,
-  theme: appearance.defaultTheme,
   user: null,
   userData: null,
   roles: [],
@@ -56,17 +55,22 @@ const initialState = {
 };
 
 class App extends Component {
-  constructor(props) {
+  constructor(props, context) {
     super(props);
 
-    this.state = initialState;
+    this.state = {
+      ...initialState,
+      theme: context.defaultTheme,
+    };
   }
 
   resetState = (callback) => {
+    let { defaultTheme } = this.context;
+
     this.setState(
       {
         ready: true,
-        theme: appearance.defaultTheme,
+        theme: defaultTheme,
         user: null,
         userData: null,
         roles: [],
@@ -75,24 +79,48 @@ class App extends Component {
     );
   };
 
-  setTheme = (theme, callback) => {
-    if (!theme) {
-      this.setState(
-        {
-          theme: appearance.defaultTheme,
-        },
-        callback
-      );
+  setTheme = (theme, updateData = true) => {
+    const { createTheme, defaultTheme, changeTheme } = this.context;
 
-      return;
-    }
+    return new Promise((resolve, reject) => {
+      if (!theme) {
+        this.setState({ theme: defaultTheme }, () => {
+          resolve();
+        });
 
-    this.setState(
-      {
-        theme: appearance.createTheme(theme),
-      },
-      callback
-    );
+        return;
+      }
+
+      this.setState({ theme: createTheme(theme) }, () => {
+        if (updateData) {
+          changeTheme(theme)
+            .then(() => {
+              resolve();
+            })
+            .catch((reason) => {
+              reject(reason);
+            });
+        } else {
+          resolve();
+        }
+      });
+    });
+  };
+
+  resetTheme = () => {
+    const { resetTheme: process, defaultTheme } = this.context;
+
+    return new Promise((resolve, reject) => {
+      process()
+        .then(() => {
+          this.setState({ theme: defaultTheme }, () => {
+            resolve();
+          });
+        })
+        .catch((reason) => {
+          reject(reason);
+        });
+    });
   };
 
   openDialog = (dialogId, callback) => {
@@ -265,150 +293,162 @@ class App extends Component {
 
     return (
       <StyledEngineProvider injectFirst>
-        <ThemeProvider theme={theme}>
-          <CssBaseline />
+        <ThemeContext.Provider
+          value={{
+            ...appearance,
+            theme,
+            setTheme: this.setTheme,
+            resetTheme: this.resetTheme,
+          }}
+        >
+          <ThemeProvider theme={theme}>
+            <CssBaseline />
 
-          <ErrorBoundary>
-            {!ready && <LaunchScreen />}
+            <ErrorBoundary>
+              {!ready && <LaunchScreen />}
 
-            {ready && (
-              <>
-                <Router
-                  user={user}
-                  roles={roles}
-                  bar={
-                    <Bar
-                      performingAction={performingAction}
-                      theme={theme}
-                      user={user}
-                      userData={userData}
-                      roles={roles}
-                      onSignUpClick={() => this.openDialog("signUpDialog")}
-                      onSignInClick={() => this.openDialog("signInDialog")}
-                      onAboutClick={() => this.openDialog("aboutDialog")}
-                      onSettingsClick={() => this.openDialog("settingsDialog")}
-                      onSignOutClick={() => this.openDialog("signOutDialog")}
-                    />
-                  }
-                  openSnackbar={this.openSnackbar}
-                />
+              {ready && (
+                <>
+                  <Router
+                    user={user}
+                    roles={roles}
+                    bar={
+                      <Bar
+                        performingAction={performingAction}
+                        theme={theme}
+                        user={user}
+                        userData={userData}
+                        roles={roles}
+                        onSignUpClick={() => this.openDialog("signUpDialog")}
+                        onSignInClick={() => this.openDialog("signInDialog")}
+                        onAboutClick={() => this.openDialog("aboutDialog")}
+                        onSettingsClick={() =>
+                          this.openDialog("settingsDialog")
+                        }
+                        onSignOutClick={() => this.openDialog("signOutDialog")}
+                      />
+                    }
+                    openSnackbar={this.openSnackbar}
+                  />
 
-                <DialogHost
-                  performingAction={performingAction}
-                  theme={theme}
-                  user={user}
-                  userData={userData}
-                  openSnackbar={this.openSnackbar}
-                  dialogs={{
-                    aboutDialog: {
-                      dialogProps: {
-                        open: aboutDialog.open,
+                  <DialogHost
+                    performingAction={performingAction}
+                    theme={theme}
+                    user={user}
+                    userData={userData}
+                    openSnackbar={this.openSnackbar}
+                    dialogs={{
+                      aboutDialog: {
+                        dialogProps: {
+                          open: aboutDialog.open,
 
-                        onClose: () => this.closeDialog("aboutDialog"),
-                      },
-                    },
-
-                    signUpDialog: {
-                      dialogProps: {
-                        open: signUpDialog.open,
-
-                        onClose: (callback) => {
-                          this.closeDialog("signUpDialog");
-
-                          if (callback && typeof callback === "function") {
-                            callback();
-                          }
+                          onClose: () => this.closeDialog("aboutDialog"),
                         },
                       },
-                    },
 
-                    signInDialog: {
-                      dialogProps: {
-                        open: signInDialog.open,
+                      signUpDialog: {
+                        dialogProps: {
+                          open: signUpDialog.open,
 
-                        onClose: (callback) => {
-                          this.closeDialog("signInDialog");
+                          onClose: (callback) => {
+                            this.closeDialog("signUpDialog");
 
-                          if (callback && typeof callback === "function") {
-                            callback();
-                          }
+                            if (callback && typeof callback === "function") {
+                              callback();
+                            }
+                          },
                         },
                       },
-                    },
 
-                    settingsDialog: {
-                      dialogProps: {
-                        open: settingsDialog.open,
+                      signInDialog: {
+                        dialogProps: {
+                          open: signInDialog.open,
 
-                        onClose: () => this.closeDialog("settingsDialog"),
+                          onClose: (callback) => {
+                            this.closeDialog("signInDialog");
+
+                            if (callback && typeof callback === "function") {
+                              callback();
+                            }
+                          },
+                        },
                       },
 
-                      props: {
-                        onDeleteAccountClick: () =>
-                          this.openDialog("deleteAccountDialog"),
-                      },
-                    },
+                      settingsDialog: {
+                        dialogProps: {
+                          open: settingsDialog.open,
 
-                    deleteAccountDialog: {
-                      dialogProps: {
-                        open: deleteAccountDialog.open,
+                          onClose: () => this.closeDialog("settingsDialog"),
+                        },
 
-                        onClose: () => this.closeDialog("deleteAccountDialog"),
-                      },
-
-                      props: {
-                        deleteAccount: this.deleteAccount,
-                      },
-                    },
-
-                    signOutDialog: {
-                      dialogProps: {
-                        open: signOutDialog.open,
-
-                        onClose: () => this.closeDialog("signOutDialog"),
+                        props: {
+                          onDeleteAccountClick: () =>
+                            this.openDialog("deleteAccountDialog"),
+                        },
                       },
 
-                      props: {
-                        title: "Sign out?",
-                        contentText:
-                          "While signed out you are unable to manage your profile and conduct other activities that require you to be signed in.",
-                        dismissiveAction: (
-                          <Button
-                            color="primary"
-                            onClick={() => this.closeDialog("signOutDialog")}
-                          >
-                            Cancel
-                          </Button>
-                        ),
-                        confirmingAction: (
-                          <Button
-                            color="primary"
-                            disabled={performingAction}
-                            variant="contained"
-                            onClick={this.signOut}
-                          >
-                            Sign Out
-                          </Button>
-                        ),
-                      },
-                    },
-                  }}
-                />
+                      deleteAccountDialog: {
+                        dialogProps: {
+                          open: deleteAccountDialog.open,
 
-                <Snackbar
-                  anchorOrigin={{
-                    vertical: "bottom",
-                    horizontal: "center",
-                  }}
-                  autoHideDuration={snackbar.autoHideDuration}
-                  message={snackbar.message}
-                  open={snackbar.open}
-                  onClose={this.closeSnackbar}
-                />
-              </>
-            )}
-          </ErrorBoundary>
-        </ThemeProvider>
+                          onClose: () =>
+                            this.closeDialog("deleteAccountDialog"),
+                        },
+
+                        props: {
+                          deleteAccount: this.deleteAccount,
+                        },
+                      },
+
+                      signOutDialog: {
+                        dialogProps: {
+                          open: signOutDialog.open,
+
+                          onClose: () => this.closeDialog("signOutDialog"),
+                        },
+
+                        props: {
+                          title: "Sign out?",
+                          contentText:
+                            "While signed out you are unable to manage your profile and conduct other activities that require you to be signed in.",
+                          dismissiveAction: (
+                            <Button
+                              color="primary"
+                              onClick={() => this.closeDialog("signOutDialog")}
+                            >
+                              Cancel
+                            </Button>
+                          ),
+                          confirmingAction: (
+                            <Button
+                              color="primary"
+                              disabled={performingAction}
+                              variant="contained"
+                              onClick={this.signOut}
+                            >
+                              Sign Out
+                            </Button>
+                          ),
+                        },
+                      },
+                    }}
+                  />
+
+                  <Snackbar
+                    anchorOrigin={{
+                      vertical: "bottom",
+                      horizontal: "center",
+                    }}
+                    autoHideDuration={snackbar.autoHideDuration}
+                    message={snackbar.message}
+                    open={snackbar.open}
+                    onClose={this.closeSnackbar}
+                  />
+                </>
+              )}
+            </ErrorBoundary>
+          </ThemeProvider>
+        </ThemeContext.Provider>
       </StyledEngineProvider>
     );
   }
@@ -443,7 +483,7 @@ class App extends Component {
               authentication
                 .getRoles()
                 .then((value) => {
-                  this.setTheme(data.theme, () => {
+                  this.setTheme(data.theme, false).then(() => {
                     this.setState({
                       ready: true,
                       user: user,
@@ -504,5 +544,7 @@ class App extends Component {
     }
   }
 }
+
+App.contextType = ThemeContext;
 
 export default App;
